refactor(components): split group limits
